Refactor: moved 'Get Module Info' to a separate function
